requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.4
dj-database-url==3.1.2
Django==6.0.3
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gunicorn==25.1.0
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
psycopg[binary]==3.2.10
PyJWT==2.11.0
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
